+++ src/ui/first_run/default_paths.rs
@@ -1,14 +1,15 @@
+use std::path::PathBuf;
+
 use relm4::prelude::*;
 use relm4::component::*;
 
 use adw::prelude::*;
 
-use std::path::PathBuf;
-
 use crate::*;
 use crate::i18n::*;
-use super::main::*;
 use crate::ui::components::progress_bar::*;
+
+use super::main::*;
 
 pub struct DefaultPathsApp {
     progress_bar: AsyncController<ProgressBar>,