+++ src/ui/preferences/general/components.rs
@@ -7,13 +7,11 @@ use adw::prelude::*;
 use anime_launcher_sdk::wincompatlib::prelude::*;
 
 use anime_launcher_sdk::components::*;
-use anime_launcher_sdk::components::wine::WincompatlibWine;
+use anime_launcher_sdk::components::wine::UnifiedWine;
 
 use super::GeneralAppMsg;
 
-use crate::ui::components;
 use crate::ui::components::*;
-
 use crate::i18n::*;
 use crate::*;
 
@@ -345,12 +343,12 @@ impl SimpleAsyncComponent for ComponentsPage {
         match msg {
             ComponentsPageMsg::WineRecommendedOnly(state) => {
                 // todo
-                self.wine_components.sender().send(components::list::AppMsg::ShowRecommendedOnly(state)).unwrap();
+                self.wine_components.sender().send(ComponentsListMsg::ShowRecommendedOnly(state)).unwrap();
             }
 
             ComponentsPageMsg::DxvkRecommendedOnly(state) => {
                 // todo
-                self.dxvk_components.sender().send(components::list::AppMsg::ShowRecommendedOnly(state)).unwrap();
+                self.dxvk_components.sender().send(ComponentsListMsg::ShowRecommendedOnly(state)).unwrap();
             }
 
             ComponentsPageMsg::UpdateDownloadedWine => {
@@ -426,7 +424,7 @@ impl SimpleAsyncComponent for ComponentsPage {
                             let wine_name = version.name.to_string();
 
                             std::thread::spawn(move || {
-                                match wine.update_prefix::<&str>(None) {
+                                match wine.update_prefix(None::<&str>) {
                                     Ok(_) => {
                                         config.game.wine.selected = Some(wine_name); 
 
@@ -463,8 +461,8 @@ impl SimpleAsyncComponent for ComponentsPage {
                                 let mut wine = match config.get_selected_wine() {
                                     Ok(Some(version)) => {
                                         match version.to_wine(config.components.path, Some(config.game.wine.builds.join(&version.name))) {
-                                            WincompatlibWine::Default(wine) => wine,
-                                            WincompatlibWine::Proton(_) => return
+                                            UnifiedWine::Default(wine) => wine,
+                                            UnifiedWine::Proton(_) => return
                                         }
                                     }