+++ CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.7.1] - 11.06.2023
+
+### Added
+
+- Added Discord RPC icon selection
+- Added Japanese
+- Added Hungarian
+
+### Fixed
+
+- Fixed progress bar style after running game repairer
+- Fixed repair button functionality (#186)
+- Fixed default launcher language selection at the first start
+- Fixed some installer updates reporting (including "checking free space")
+
+### Changed
+
+- Reworked game sessions selection
+- Updated Indonesian
+- Updated French
+- Made initial tasks async which has decreased startup time
+- Updated fps unlocker to 2.1.1
+
+### Removed
+
+- Removed 3.7.0 workaround
+- Removed patch mirror migration
+
 ## [3.7.0] - 24.05.2023
 
 ### Added
@@ -264,7 +292,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <br>
 
-[unreleased]: https://github.com/an-anime-team/an-anime-game-launcher/compare/3.7.0...next
+[unreleased]: https://github.com/an-anime-team/an-anime-game-launcher/compare/3.7.1...next
+[3.7.1]: https://github.com/an-anime-team/an-anime-game-launcher/compare/3.7.0...3.7.1
 [3.7.0]: https://github.com/an-anime-team/an-anime-game-launcher/compare/3.6.0...3.7.0
 [3.6.0]: https://github.com/an-anime-team/an-anime-game-launcher/compare/3.5.2...3.6.0
 [3.5.2]: https://github.com/an-anime-team/an-anime-game-launcher/compare/3.5.1...3.5.2